test: Cleanup code
